Switch back to per job fixtures
They are way easier to handle this way as it seems. We might need to change stuff if we every get slow performance from dropping the database and i really hope that it works on travis this time as I have been having weird issues in the last few pushes.